Create 'deployEnabled' parameter to control activation of deployments
Allow the deployment of disabled WARs by using the 'add' functionality
of JBoss AS in place of 'deploy'.

The default state is enabled, i.e. the application will be deployed and
launched in a single action. By setting deployEnabled to false, the
content will be uploaded to the Application Server, but remain disabled
until manually started through the management interface.